requirements.txt
psycopg2
django
djangorestframework
djangorestframework-jwt